Refactor home page to "my decks"
